Edits to the following to add the functionality to apply a best fit W plane from some time in the future.
This improves the execution time for long tracks by close to a factor of 2 as the imaging time is dominated by grinding and degridding and this reduces the update rate of the w-plane by a factor of 2.

This involved:

Changes to:

Code/Base/accessors/current/dataaccess/BestWPlaneDataAccessor.cc
Code/Base/accessors/current/dataaccess/BestWPlaneDataAccessor.h

Which were the addition of a new protected member function 

double updateAdvancedTimePlaneIfNecessary(double tolerance, const casa::MDirection &tangentPoint) const;

And a public set function with two private variables a bool that turns this feature on and off and set the 
update rate.


In order to implement this I had to change the interface to the SnapshotGriddingAdapter. I added a member bool and an extra parameter to the constructor to set it. This is involves changing the VisGridderFactory to read the parset and find the flag to set it in the constructor if necessary. So small changes to:

Code/Components/Synthesis/synthesis/current/gridding/SnapShotImagingGridderAdapter.cc
Code/Components/Synthesis/synthesis/current/gridding/SnapShotImagingGridderAdapter.h
Code/Components/Synthesis/synthesis/current/gridding/VisGridderFactory.cc

I also updated the documentation in:

Code/Doc/user/current/doc/calim/gridder.rst