ASKAPSDP-2126 - Improving cbpcalibrator run-time.
This change moves the SolutionAccessor up a level from the 
writeModel() function to the run() function, making it a 
BPCalibratorParallel member. This allows us to hold off writing the 
table entries to disk until all are compiled on the master, greatly 
improving the run time as we don't have lots of little I/O calls.