Including Coordinate.h to fix bug in casacore
I'm about to open a pull request for this one: casacore's ImageRegrid.h
forward-defines the Coordinate class, but in fact it actually requires
it fully defined because an inner member is inspected at some point
(i.e., Coordinate::Type in ImageRegrid.h:209). This error seems to be
present in casacore 3 too.

To easily circumvent this we can simply include the full definition of
the class, which is probably why this hasn't been apparent until now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>