Dumping out convolution correction factors matrix
Also, to correctly align the other set of outputs we already produce we
are passing down the iteration number. This required a number of changes
to a few gridder classes, but we don't really mean to have these changes
incorporated into yandasoft anyway, or at least not in this shape.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>