Target-based dependencies and exporting
This one commit greatly modifies the existing cmake infrastructure of
this project. Firstly, it makes sure that all compilation settings are
set at the target level (rather than at the directory level), allowing
them to be used by target dependencies transitively. To simplify things
all (or almost all) properties have been made PUBLIC, meaning that
dependencies will see these properties and use them when going through
compilation.

Secondly, we introduce an exporting mechanism in which we generate a
project configuration file (that can be loaded by find_package()), we
place it under the project's binary directory (and export the project's
binary directory as a package), and also make sure it gets copied into
the package's installation directory. Separately we also generate a
targets file with our targets information, and also place those under
the project's binary directory and under the package's installation
directory. These changes make projects searchable via find_package(),
both when installed into an installation area, and when included as a
sub-project of a parent cmake project (i.e., via add_subdirectory). With
these changes the FindXYZ.cmake modules in askap-cmake are not
deprecated and not needed.

A small, final change is *where* the askap-cmake module is searched for:
by switching from CMAKE_CURRENT_SOURCE_DIR to CMAKE_SOURCE_DIR we enable
a single copy of askap-cmake when compiling repositories as part of a
single, bigger project (which might already have a copy of askap-cmake
anyway).

These changes enable the compilation of many projects at once, or
separately. This in turn removes the need for automatic compilation of
missing askap dependencies (which was error prone), allows for better
development environments and IDE integration, and promotes better
synchronisation between the different repositories (e.g., if a parent
repository with git submodules drives the overall compilation).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>