Timing things more accurately
The times reported for "calculated normal equation" have actually always
included the creation of the measurement equation and the accumulation
of data during the pre-averaging step, which is usually a one off; hence
the first major iteration always reported more time than the other ones.

With this all iterations should report similar times. Making up for the
"lost" reported time we added a new timer around the data accumulation
step, which is what takes most of the time, and report that separately.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>