Replace "myAllocationStart" by "channel"
The "myAllocationStart" variable is only used within the channels loop,
and effectively indicates the channel that is being processed on each
iteration. Thus, it makes sense to define it only within the scope of
the loop, and to give it the more meaningful "channel" name to make its
intention clearer.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>