Improve channel allocation strategy and looping
The previous version of the code had two problems: firstly, the channel
allocation strategy was not optimal, as all remainder channels were
allocated to the last rank. Thus, in the worst-case scenario the last
rank had to process (rank - 1) extra channels. As an example, this was
the allocation with 11 channels and 4 ranks:

Rank 0 - RankAllocation starts at 0 and is 2 in size
Rank 1 - RankAllocation starts at 2 and is 2 in size
Rank 2 - RankAllocation starts at 4 and is 2 in size
Rank 3 - RankAllocation starts at 6 and is 5 in size

The second problem was that the code adjusting the load on the last rank
did not update the value of myFullAllocationStop. This had the
unintended consequence that when one used less ranks than total number
of channels, the last ranks had 0 channels to process, but still entered
the body of the channel loop. This in turn caused exceptions in the loop
body to be raised in these ranks, which led to skipping the call to
comms.barrier (and to the premature destruction of the AskapParallel
object, which calls MPI_Finalize), causing a stall on the other ranks,
effectively blocking the entire program.

This commit introduces fixes for these two problems. The first problem
is solved by using a round-robin strategy to allocate channels to ranks,
where `rem` extra channels are allocated one to each of the first `rem`
ranks. In the example above this is the new allocation:

Rank 0 - RankAllocation starts at 2 and is 3 in size
Rank 1 - RankAllocation starts at 4 and is 3 in size
Rank 2 - RankAllocation starts at 6 and is 3 in size
Rank 3 - RankAllocation starts at 9 and is 2 in size

The second problem is fixed by removing the usage of the extra
myFullAllocationStop variable, and using the start/size information
instead directly in the loop condition. This avoids ranks with no work
from entering the loop, avoiding exceptions, and therefore avoiding not
hitting the barrier.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>